+++ CumulusMX/Cumulus.cs
@@ -10190,10 +10190,13 @@ namespace CumulusMX
 				{
 					_ = Task.Run(async () =>
 					{
+						LogDebugMessage($"ProcessHttpFiles: Downloading from {item.Url} to {item.Remote}");
+
 						// just create a file from the download
 						await httpFiles.DownloadHttpFile(item.Url, item.Remote);
-
 						item.SetNextInterval(now);
+
+						LogDebugMessage($"ProcessHttpFiles: Downloading from {item.Url} to {item.Remote}");
 					});
 				}
 
@@ -10228,27 +10231,25 @@ namespace CumulusMX
 
 					foreach (var item in uploads)
 					{
-						Stream strm = null;
-
 						try
 						{
-
 							if (Program.ExitSystemToken.IsCancellationRequested)
 								return;
 
-							strm = httpFiles.DownloadHttpFileStream(item.Url).Result;
-							UploadStream(conn, item.Remote, strm, -1);
+							LogDebugMessage($"ProcessHttpFiles: Downloading from {item.Url}, upload to {item.Remote}");
 
+							using var strm = await httpFiles.DownloadFileStream(item.Url);
+							if (strm != null)
+							{
+								UploadStream(conn, item.Remote, strm, -1);
+							}
 							item.SetNextInterval(now);
+
+							LogDebugMessage($"ProcessHttpFiles: Download from {item.Url}, upload to {item.Remote} complete");
 						}
 						catch (Exception ex) when (ex is not TaskCanceledException)
 						{
 							LogExceptionMessage(ex, $"ProcessHttpFiles: Error uploading http file {item.Url} to: {item.Remote}");
-						}
-						finally
-						{
-							if (null != strm)
-								strm.Dispose();
 						}
 					}
 
@@ -10311,27 +10312,26 @@ namespace CumulusMX
 
 				foreach (var item in uploads)
 				{
-					Stream strm = null;
-
 					try
 					{
 
 						if (Program.ExitSystemToken.IsCancellationRequested)
 							return;
 
-						strm = httpFiles.DownloadHttpFileStream(item.Url).Result;
-						await UploadStream(conn, item.Remote, strm, -1);
+						LogDebugMessage($"ProcessHttpFiles: Downloading from {item.Url}, upload to {item.Remote}");
 
+						using var strm = await httpFiles.DownloadFileStream(item.Url);
+						if (strm != null)
+						{
+							await UploadStream(conn, item.Remote, strm, -1);
+						}
 						item.SetNextInterval(now);
+
+						LogDebugMessage($"ProcessHttpFiles: Download from {item.Url}, upload to {item.Remote} complete");
 					}
 					catch (Exception ex) when (ex is not TaskCanceledException)
 					{
 						LogExceptionMessage(ex, $"ProcessHttpFiles: Error uploading http file {item.Url} to: {item.Remote}");
-					}
-					finally
-					{
-						if (null != strm)
-							strm.Dispose();
 					}
 				}
 
@@ -10387,10 +10387,12 @@ namespace CumulusMX
 							if (Program.ExitSystemToken.IsCancellationRequested)
 								return false;
 
+							LogDebugMessage($"ProcessHttpFiles: Uploading http file {downloadfile} to: {remotefile}");
 							var content = await httpFiles.DownloadHttpFileBase64String(item.Url);
 							await UploadString(phpUploadHttpClient, false, null, content, item.Remote, -1, true);
 
 							item.SetNextInterval(now);
+							LogDebugMessage($"ProcessHttpFiles: Upload of http file {downloadfile} to: {remotefile} complete");
 						}
 						catch (Exception ex) when (ex is not TaskCanceledException)
 						{
@@ -10561,14 +10563,14 @@ namespace CumulusMX
 									// have we already uploaded the base file?
 									if (item.logFileLastLineNumber > 0)
 									{
-										if (AppendText(conn, remotefile, data, -1, linesAdded))
+										if (AppendText(conn, remotefile, data, cycle1k, linesAdded))
 										{
 											ActiveExtraFiles[i].logFileLastLineNumber += linesAdded;
 										}
 									}
 									else // no, just upload the base file
 									{
-										if (UploadFile(conn, uploadfile, remotefile, -1))
+										if (UploadFile(conn, uploadfile, remotefile, cycle1k))
 										{
 											ActiveExtraFiles[i].logFileLastLineNumber += linesAdded;
 										}
@@ -10702,7 +10704,7 @@ namespace CumulusMX
 							try
 							{
 								LogDebugMessage($"{msgPrefix} Uploading Moon image file");
-								if (UploadFile(conn, Path.Combine("web", "moon.png"), remotePath + MoonImage.FtpDest, -1))
+								if (UploadFile(conn, Path.Combine("web", "moon.png"), remotePath + MoonImage.FtpDest, cycle1k))
 								{
 									// clear the image ready for FTP flag, only upload once an hour
 									MoonImage.ReadyToFtp = false;
@@ -10860,7 +10862,7 @@ namespace CumulusMX
 								// have we already uploaded the base file?
 								if (item.logFileLastLineNumber > 0)
 								{
-									if (await AppendText(conn, remotefile, data, -1, linesAdded))
+									if (await AppendText(conn, remotefile, data, cycle1k, linesAdded))
 									{
 										ActiveExtraFiles[i].logFileLastLineNumber += linesAdded;
 									}
@@ -11645,21 +11647,21 @@ namespace CumulusMX
 		// Return False if the connection is disposed, null, or not connected
 		private async Task<bool> UploadFile(AsyncFtpClient conn, string localfile, string remotefile, int cycle)
 		{
-			string cycleStr;
+			string prefix;
 			bool realtime;
 			if (cycle == 9999)
 			{
-				cycleStr = "NOAA";
+				prefix = "FTP[NOAA]";
 				realtime = false;
 			}
 			else if (cycle >= 1000)
 			{
-				cycleStr = "Int-" + (cycle - 1000);
+				prefix = $"FTP[Int-{cycle - 1000}]";
 				realtime = false;
 			}
 			else
 			{
-				cycleStr = cycle.ToString();
+				prefix = $"FTP[{cycle}]";
 				realtime = true;
 			}
 
@@ -11669,7 +11671,7 @@ namespace CumulusMX
 			{
 				if (!File.Exists(localfile))
 				{
-					LogWarningMessage($"FTP[{cycleStr}]: Error! Local file not found, aborting upload: {localfile}");
+					LogWarningMessage($"{prefix}: Error! Local file not found, aborting upload: {localfile}");
 					FtpAlarm.LastMessage = $"Error! Local file not found, aborting upload: {localfile}";
 					FtpAlarm.Triggered = true;
 					return true;
@@ -11680,14 +11682,14 @@ namespace CumulusMX
 			}
 			catch (Exception ex)
 			{
-				LogFtpMessage($"FTP[{cycleStr}]: Error {localfile} - {ex.Message}", realtime);
+				LogFtpMessage($"{prefix}: Error {localfile} - {ex.Message}", realtime);
 				FtpAlarm.LastMessage = $"Error {localfile} - {ex.Message}";
 				FtpAlarm.Triggered = true;
 
 				if (ex.InnerException != null)
 				{
-					LogFtpMessage($"FTP[{cycleStr}]: Inner Exception: {ex.GetBaseException().Message}", realtime);
-					LogExceptionMessage(ex, $"FTP[{cycleStr}]: Error {localfile}");
+					LogFtpMessage($"{prefix}: Inner Exception: {ex.GetBaseException().Message}", realtime);
+					LogExceptionMessage(ex, $"{prefix}: Error {localfile}");
 				}
 			}
 
@@ -11696,19 +11698,19 @@ namespace CumulusMX
 
 		private bool UploadFile(SftpClient conn, string localfile, string remotefile, int cycle = -1)
 		{
-			string cycleStr;
+			string prefix;
 			if (cycle == 9999)
 			{
-				cycleStr = "NOAA";
+				prefix = "SFTP[NOAA]";
 			}
 			else
 			{
-				cycleStr = cycle >= 1000 ? "Int-" + (cycle - 1000) : (cycle.ToString());
+				prefix = $"SFTP[{(cycle >= 1000 ? $"Int - {cycle - 1000}" : cycle)}]";
 			}
 
 			if (!File.Exists(localfile))
 			{
-				LogWarningMessage($"SFTP[{cycleStr}]: Error! Local file not found, aborting upload: {localfile}");
+				LogWarningMessage($"{prefix}: Error! Local file not found, aborting upload: {localfile}");
 				FtpAlarm.LastMessage = $"Error! Local file not found, aborting upload: {localfile}";
 				FtpAlarm.Triggered = true;
 
@@ -11719,14 +11721,14 @@ namespace CumulusMX
 			{
 				if (conn == null || !conn.IsConnected)
 				{
-					LogMessage($"SFTP[{cycleStr}]: The SFTP object is null or not connected - skipping upload of {localfile}");
+					LogMessage($"{prefix}: The SFTP object is null or not connected - skipping upload of {localfile}");
 
 					return false;
 				}
 			}
 			catch (ObjectDisposedException)
 			{
-				LogErrorMessage($"SFTP[{cycleStr}]: The SFTP object is disposed - skipping upload of {localfile}");
+				LogErrorMessage($"{prefix}: The SFTP object is disposed - skipping upload of {localfile}");
 				FtpAlarm.LastMessage = $"The SFTP object is disposed - skipping upload of {localfile}";
 				FtpAlarm.Triggered = true;
 
@@ -11740,7 +11742,7 @@ namespace CumulusMX
 			}
 			catch (Exception ex)
 			{
-				LogWarningMessage($"SFTP[{cycleStr}]: Error reading {localfile} - {ex.Message}");
+				LogWarningMessage($"{prefix}: Error reading {localfile} - {ex.Message}");
 
 				FtpAlarm.LastMessage = $"Error reading {localfile} - {ex.Message}";
 				FtpAlarm.Triggered = true;
@@ -11748,7 +11750,7 @@ namespace CumulusMX
 				if (ex.InnerException != null)
 				{
 					ex = Utils.GetOriginalException(ex);
-					LogDebugMessage($"SFTP[{cycleStr}]: Base exception - {ex.Message}");
+					LogDebugMessage($"{prefix}: Base exception - {ex.Message}");
 				}
 
 			}
@@ -11765,6 +11767,10 @@ namespace CumulusMX
 			else if (cycle >= 1000)
 			{
 				prefix = $"PHP[Int-{cycle - 1000}]";
+			}
+			else if (cycle == -1)
+			{
+				prefix = "PHP[HttpFiles]";
 			}
 			else
 			{
@@ -11827,29 +11833,34 @@ namespace CumulusMX
 		private async Task<bool> UploadStream(AsyncFtpClient conn, string remotefile, Stream dataStream, int cycle)
 		{
 			string remotefiletmp = FTPRename ? remotefile + "tmp" : remotefile;
-			string cycleStr;
+			string prefix;
 			bool realtime;
 			if (cycle == 9999)
 			{
-				cycleStr = "NOAA";
+				prefix = "FTP[NOAA]";
 				realtime = false;
 			}
 			else if (cycle >= 1000)
 			{
-				cycleStr = "Int-" + (cycle - 1000);
+				prefix = "FTP[Int-{(cycle - 1000)}]";
+				realtime = false;
+			}
+			else if (cycle == -1)
+			{
+				prefix = "FTP[HttpFiles]";
 				realtime = false;
 			}
 			else
 			{
-				cycleStr = cycle.ToString();
+				prefix = "FTP[{cycle}]";
 				realtime = true;
 			}
 
 			try
 			{
-				if (dataStream.Length == 0)
+				if (dataStream.CanSeek && dataStream.Length == 0)
 				{
-					LogWarningMessage($"FTP[{cycleStr}]: The data is empty - skipping upload of {remotefile}");
+					LogWarningMessage($"{prefix}: The data is empty - skipping upload of {remotefile}");
 					FtpAlarm.LastMessage = $"The data is empty - skipping upload of {remotefile}";
 					FtpAlarm.Triggered = true;
 
@@ -11858,7 +11869,7 @@ namespace CumulusMX
 
 				if (!conn.IsConnected)
 				{
-					LogDebugMessage($"FTP[{cycleStr}]: Not connected, skipping upload of {remotefile}");
+					LogDebugMessage($"{prefix}: Not connected, skipping upload of {remotefile}");
 					return false;
 				}
 
@@ -11874,34 +11885,34 @@ namespace CumulusMX
 					}
 					catch
 					{
-						LogDebugMessage($"FTP[{cycleStr}]: Not connected after tmp file check, skipping upload of {remotefile}");
+						LogDebugMessage($"{prefix}: Not connected after tmp file check, skipping upload of {remotefile}");
 						return false;
 					}
 				}
 
-				LogFtpDebugMessage($"FTP[{cycleStr}]: Uploading {remotefiletmp}", realtime);
+				LogFtpDebugMessage($"{prefix}: Uploading {remotefiletmp}", realtime);
 
 				FtpStatus status;
 				status = await conn.UploadStream(dataStream, remotefiletmp, DeleteBeforeUpload ? FtpRemoteExists.Overwrite : FtpRemoteExists.NoCheck, false, null, Program.ExitSystemToken);
 
 				if (status.IsFailure())
 				{
-					LogErrorMessage($"FTP[{cycleStr}]: Upload of {remotefile} failed");
+					LogErrorMessage($"{prefix}: Upload of {remotefile} failed");
 					return false;
 				}
 				else if (FTPRename)
 				{
 					// rename the file
-					LogFtpDebugMessage($"FTP[{cycleStr}]: Renaming {remotefiletmp} to {remotefile}", realtime);
+					LogFtpDebugMessage($"{prefix}: Renaming {remotefiletmp} to {remotefile}", realtime);
 
 					try
 					{
 						await conn.Rename(remotefiletmp, remotefile, Program.ExitSystemToken);
-						LogFtpDebugMessage($"FTP[{cycleStr}]: Renamed {remotefiletmp}", realtime);
+						LogFtpDebugMessage($"{prefix}: Renamed {remotefiletmp}", realtime);
 					}
 					catch (Exception ex)
 					{
-						LogFtpMessage($"FTP[{cycleStr}]: Error renaming {remotefiletmp} to {remotefile} : {ex.Message}", realtime);
+						LogFtpMessage($"{prefix}: Error renaming {remotefiletmp} to {remotefile} : {ex.Message}", realtime);
 
 						FtpAlarm.LastMessage = $"Error renaming {remotefiletmp} to {remotefile} : {ex.Message}";
 						FtpAlarm.Triggered = true;
@@ -11909,7 +11920,7 @@ namespace CumulusMX
 						if (ex.InnerException != null)
 						{
 							ex = Utils.GetOriginalException(ex);
-							LogFtpMessage($"FTP[{cycleStr}]: Base exception - {ex.Message}", realtime);
+							LogFtpMessage($"{prefix}: Base exception - {ex.Message}", realtime);
 						}
 
 						return false;
@@ -11918,12 +11929,12 @@ namespace CumulusMX
 			}
 			catch (Exception ex)
 			{
-				LogFtpMessage($"FTP[{cycleStr}]: Error uploading {remotefile} : {ex.Message}", realtime);
+				LogFtpMessage($"{prefix}: Error uploading {remotefile} : {ex.Message}", realtime);
 
 				if (ex.InnerException != null)
 				{
-					LogFtpMessage($"FTP[{cycleStr}]: Inner Exception: {ex.GetBaseException().Message}", realtime);
-					LogExceptionMessage(ex, $"FTP[{cycleStr}]: Exception dump", false);
+					LogFtpMessage($"{prefix}: Inner Exception: {ex.GetBaseException().Message}", realtime);
+					LogExceptionMessage(ex, $"{prefix}: Exception dump", false);
 				}
 
 
@@ -11941,12 +11952,29 @@ namespace CumulusMX
 		private bool UploadStream(SftpClient conn, string remotefile, Stream dataStream, int cycle)
 		{
 			string remotefilename = FTPRename ? remotefile + "tmp" : remotefile;
-			string cycleStr = cycle >= 1000 ? "Int-" + (cycle - 1000) : cycle.ToString();
+			string prefix;
+			if (cycle == 9999)
+			{
+				prefix = "SFTP[NOAA]";
+			}
+			else if (cycle >= 1000)
+			{
+				prefix = $"SFTP[Int-{cycle - 1000}]";
+			}
+			else if (cycle == -1)
+			{
+				prefix = "SFTP[HttpFiles]";
+			}
+			else
+			{
+				prefix = $"SFTP[{cycle}]";
+			}
+
 			bool realtime = cycle < 1000;
 
-			if (dataStream.Length == 0)
+			if (dataStream.CanSeek && dataStream.Length == 0)
 			{
-				LogWarningMessage($"SFTP[{cycleStr}]: The data is empty - skipping upload of {remotefile}");
+				LogWarningMessage($"{prefix}: The data is empty - skipping upload of {remotefile}");
 				FtpAlarm.LastMessage = $"The data is empty - skipping upload of {remotefile}";
 				FtpAlarm.Triggered = true;
 				return true;
@@ -11954,7 +11982,7 @@ namespace CumulusMX
 
 			if (!conn.IsConnected)
 			{
-				LogDebugMessage($"SFTP[{cycleStr}]: Not connected, skipping upload of {remotefile}");
+				LogDebugMessage($"{prefix}: Not connected, skipping upload of {remotefile}");
 				return false;
 			}
 
@@ -11962,7 +11990,7 @@ namespace CumulusMX
 			{
 				if (conn == null || !conn.IsConnected)
 				{
-					LogErrorMessage($"SFTP[{cycleStr}]: The SFTP object is null or not connected - skipping upload of {remotefile}");
+					LogErrorMessage($"{prefix}: The SFTP object is null or not connected - skipping upload of {remotefile}");
 					FtpAlarm.LastMessage = $"The SFTP object is null or not connected - skipping upload of {remotefile}";
 					FtpAlarm.Triggered = true;
 
@@ -11971,7 +11999,7 @@ namespace CumulusMX
 			}
 			catch (ObjectDisposedException)
 			{
-				LogErrorMessage($"SFTP[{cycleStr}]: The SFTP object is disposed - skipping upload of {remotefile}");
+				LogErrorMessage($"{prefix}: The SFTP object is disposed - skipping upload of {remotefile}");
 
 				FtpAlarm.LastMessage = $"The SFTP object is disposed - skipping upload of {remotefile}";
 				FtpAlarm.Triggered = true;
@@ -11984,24 +12012,24 @@ namespace CumulusMX
 				// No delete before upload required for SFTP as we use the overwrite flag
 				try
 				{
-					LogDebugMessage($"SFTP[{cycleStr}]: Uploading {remotefilename}");
+					LogDebugMessage($"{prefix}: Uploading {remotefilename}");
 
 					conn.OperationTimeout = TimeSpan.FromSeconds(15);
 					conn.UploadFile(dataStream, remotefilename); // defaults to CreateNewOrOpen
 					dataStream.Close();
 
-					LogDebugMessage($"SFTP[{cycleStr}]: Uploaded {remotefilename}");
+					LogDebugMessage($"{prefix}: Uploaded {remotefilename}");
 				}
 				catch (ObjectDisposedException)
 				{
-					LogErrorMessage($"SFTP[{cycleStr}]: The SFTP object is disposed");
+					LogErrorMessage($"{prefix}: The SFTP object is disposed");
 					FtpAlarm.LastMessage = $"The SFTP object is disposed - skipping upload of {remotefile}";
 					FtpAlarm.Triggered = true;
 					return false;
 				}
 				catch (Exception ex)
 				{
-					LogErrorMessage($"SFTP[{cycleStr}]: Error uploading {remotefilename} : {ex.Message}");
+					LogErrorMessage($"{prefix}: Error uploading {remotefilename} : {ex.Message}");
 
 					FtpAlarm.LastMessage = $"Error uploading {remotefilename} : {ex.Message}";
 					FtpAlarm.Triggered = true;
@@ -12012,7 +12040,7 @@ namespace CumulusMX
 					if (ex.InnerException != null)
 					{
 						ex = Utils.GetOriginalException(ex);
-						LogFtpMessage($"FTP[{cycleStr}]: Base exception - {ex.Message}", realtime);
+						LogFtpMessage($"{prefix}: Base exception - {ex.Message}", realtime);
 					}
 
 					// Lets start again anyway! Too hard to tell if the error is recoverable
@@ -12024,20 +12052,20 @@ namespace CumulusMX
 					// rename the file
 					try
 					{
-						LogDebugMessage($"SFTP[{cycleStr}]: Renaming {remotefilename} to {remotefile}");
+						LogDebugMessage($"{prefix}: Renaming {remotefilename} to {remotefile}");
 						conn.RenameFile(remotefilename, remotefile, true);
-						LogDebugMessage($"SFTP[{cycleStr}]: Renamed {remotefilename}");
+						LogDebugMessage($"{prefix}: Renamed {remotefilename}");
 					}
 					catch (ObjectDisposedException)
 					{
-						LogErrorMessage($"SFTP[{cycleStr}]: The SFTP object is disposed");
+						LogErrorMessage($"{prefix}: The SFTP object is disposed");
 						FtpAlarm.LastMessage = $"The SFTP object is disposed during renaming of {remotefile}";
 						FtpAlarm.Triggered = true;
 						return false;
 					}
 					catch (Exception ex)
 					{
-						LogErrorMessage($"SFTP[{cycleStr}]: Error renaming {remotefilename} to {remotefile} : {ex.Message}");
+						LogErrorMessage($"{prefix}: Error renaming {remotefilename} to {remotefile} : {ex.Message}");
 
 						FtpAlarm.LastMessage = $"Error renaming {remotefilename} to {remotefile} : {ex.Message}";
 						FtpAlarm.Triggered = true;
@@ -12045,24 +12073,24 @@ namespace CumulusMX
 						if (ex.InnerException != null)
 						{
 							ex = Utils.GetOriginalException(ex);
-							LogFtpMessage($"SFTP[{cycleStr}]: Base exception - {ex.Message}", realtime);
+							LogFtpMessage($"{prefix}: Base exception - {ex.Message}", realtime);
 						}
 
 						return false;
 					}
 				}
-				LogDebugMessage($"SFTP[{cycleStr}]: Completed uploading {remotefile}");
+				LogDebugMessage($"{prefix}: Completed uploading {remotefile}");
 			}
 			catch (ObjectDisposedException)
 			{
-				LogErrorMessage($"SFTP[{cycleStr}]: The SFTP object is disposed");
+				LogErrorMessage($"{prefix}: The SFTP object is disposed");
 				FtpAlarm.LastMessage = "The SFTP object is disposed";
 				FtpAlarm.Triggered = true;
 				return false;
 			}
 			catch (Exception ex)
 			{
-				LogErrorMessage($"SFTP[{cycleStr}]: Error uploading {remotefile} - {ex.Message}");
+				LogErrorMessage($"{prefix}: Error uploading {remotefile} - {ex.Message}");
 
 				FtpAlarm.LastMessage = $"Error uploading {remotefile} - {ex.Message}";
 				FtpAlarm.Triggered = true;
@@ -12070,7 +12098,7 @@ namespace CumulusMX
 				if (ex.InnerException != null)
 				{
 					ex = Utils.GetOriginalException(ex);
-					LogDebugMessage($"SFTP[{cycleStr}]: Base exception - {ex.Message}");
+					LogDebugMessage($"{prefix}: Base exception - {ex.Message}");
 				}
 
 				return false;
@@ -12090,11 +12118,19 @@ namespace CumulusMX
 
 		private bool AppendText(SftpClient conn, string remotefile, string text, int cycle, int linesadded)
 		{
-			string cycleStr = cycle >= 0 ? cycle.ToString() : "Int";
+			string prefix;
+			if (cycle < 1000)
+			{
+				prefix = $"FTP[{cycle}]";
+			}
+			else
+			{
+				prefix = $"FTP[Int-{cycle - 1000}]";
+			}
 
 			if (text.Length == 0)
 			{
-				LogWarningMessage($"SFTP[{cycleStr}]: The data is empty - skipping upload of {remotefile}");
+				LogWarningMessage($"{prefix}: The data is empty - skipping upload of {remotefile}");
 				FtpAlarm.LastMessage = $"The data is empty - skipping upload of {remotefile}";
 				FtpAlarm.Triggered = true;
 				return false;
@@ -12104,7 +12140,7 @@ namespace CumulusMX
 			{
 				if (conn == null || !conn.IsConnected)
 				{
-					LogErrorMessage($"SFTP[{cycleStr}]: The SFTP object is null or not connected - skipping upload of {remotefile}");
+					LogErrorMessage($"{prefix}: The SFTP object is null or not connected - skipping upload of {remotefile}");
 					FtpAlarm.LastMessage = $"The SFTP object is null or not connected - skipping upload of {remotefile}";
 					FtpAlarm.Triggered = true;
 
@@ -12113,7 +12149,7 @@ namespace CumulusMX
 			}
 			catch (ObjectDisposedException)
 			{
-				LogErrorMessage($"SFTP[{cycleStr}]: The SFTP object is disposed - skipping upload of {remotefile}");
+				LogErrorMessage($"{prefix}: The SFTP object is disposed - skipping upload of {remotefile}");
 
 				FtpAlarm.LastMessage = $"The SFTP object is disposed - skipping upload of {remotefile}";
 				FtpAlarm.Triggered = true;
@@ -12123,23 +12159,23 @@ namespace CumulusMX
 
 			try
 			{
-				LogDebugMessage($"SFTP[{cycleStr}]: Uploading {remotefile} [adding {linesadded} lines]");
+				LogDebugMessage($"{prefix}: Uploading {remotefile} [adding {linesadded} lines]");
 
 				conn.OperationTimeout = TimeSpan.FromSeconds(15);
 				conn.AppendAllText(remotefile, text);
 
-				LogDebugMessage($"SFTP[{cycleStr}]: Uploaded {remotefile} [added {linesadded} lines]");
+				LogDebugMessage($"{prefix}: Uploaded {remotefile} [added {linesadded} lines]");
 			}
 			catch (ObjectDisposedException)
 			{
-				LogErrorMessage($"SFTP[{cycleStr}]: The SFTP object is disposed");
+				LogErrorMessage($"{prefix}: The SFTP object is disposed");
 				FtpAlarm.LastMessage = $"The SFTP object is disposed - skipping upload of {remotefile}";
 				FtpAlarm.Triggered = true;
 				return false;
 			}
 			catch (Exception ex)
 			{
-				LogErrorMessage($"SFTP[{cycleStr}]: Error uploading {remotefile} : {ex.Message}");
+				LogErrorMessage($"{prefix}: Error uploading {remotefile} : {ex.Message}");
 
 				FtpAlarm.LastMessage = $"Error uploading {remotefile} : {ex.Message}";
 				FtpAlarm.Triggered = true;
@@ -12150,7 +12186,7 @@ namespace CumulusMX
 				if (ex.InnerException != null)
 				{
 					ex = Utils.GetOriginalException(ex);
-					LogDebugMessage($"FTP[{cycleStr}]: Base exception - {ex.Message}");
+					LogDebugMessage($"{prefix}: Base exception - {ex.Message}");
 				}
 
 				// Lets start again anyway! Too hard to tell if the error is recoverable
@@ -12163,12 +12199,22 @@ namespace CumulusMX
 
 		private async Task<bool> AppendText(AsyncFtpClient conn, string remotefile, string text, int cycle, int linesadded)
 		{
-			string cycleStr = cycle >= 0 ? cycle.ToString() : "Int";
-			bool realtime = cycle >= 0;
+			string prefix;
+			bool realtime;
+			if (cycle < 1000)
+			{
+				prefix = $"FTP[{cycle}]";
+				realtime = true;
+			}
+			else
+			{
+				prefix = $"FTP[Int-{cycle - 1000}]";
+				realtime = false;
+			}
 
 			if (text.Length == 0)
 			{
-				LogFtpMessage($"FTP[{cycleStr}]: The data is empty - skipping upload of {remotefile}", realtime);
+				LogFtpMessage($"{prefix}: The data is empty - skipping upload of {remotefile}", realtime);
 				FtpAlarm.LastMessage = $"The data is empty - skipping upload of {remotefile}";
 				FtpAlarm.Triggered = true;
 				return false;
@@ -12178,7 +12224,7 @@ namespace CumulusMX
 			{
 				if (conn == null || !conn.IsConnected)
 				{
-					LogFtpMessage($"FTP[{cycleStr}]: The FTP object is null or not connected - skipping upload of {remotefile}", realtime);
+					LogFtpMessage($"{prefix}: The FTP object is null or not connected - skipping upload of {remotefile}", realtime);
 					FtpAlarm.LastMessage = $"The FTP object is null or not connected - skipping upload of {remotefile}";
 					FtpAlarm.Triggered = true;
 
@@ -12187,7 +12233,7 @@ namespace CumulusMX
 			}
 			catch (ObjectDisposedException)
 			{
-				LogFtpMessage($"FTP[{cycleStr}]: The FTP object is disposed - skipping upload of {remotefile}", realtime);
+				LogFtpMessage($"FTP[{prefix}]: The FTP object is disposed - skipping upload of {remotefile}", realtime);
 
 				FtpAlarm.LastMessage = $"The FTP object is disposed - skipping upload of {remotefile}";
 				FtpAlarm.Triggered = true;
@@ -12197,22 +12243,22 @@ namespace CumulusMX
 
 			try
 			{
-				LogFtpDebugMessage($"FTP[{cycleStr}]: Uploading {remotefile} [adding {linesadded} lines]", realtime);
+				LogFtpDebugMessage($"{prefix}: Uploading {remotefile} [adding {linesadded} lines]", realtime);
 
 				await conn.UploadStream(GenerateStreamFromString(text), remotefile, FtpRemoteExists.AddToEnd, false, null, Program.ExitSystemToken);
 
-				LogFtpDebugMessage($"FTP[{cycleStr}]: Uploaded {remotefile} [added {linesadded} lines]", realtime);
+				LogFtpDebugMessage($"{prefix}: Uploaded {remotefile} [added {linesadded} lines]", realtime);
 			}
 			catch (ObjectDisposedException)
 			{
-				LogFtpMessage($"FTP[{cycleStr}]: The FTP object is disposed", realtime);
+				LogFtpMessage($"{prefix}: The FTP object is disposed", realtime);
 				FtpAlarm.LastMessage = $"The FTP object is disposed - skipping upload of {remotefile}";
 				FtpAlarm.Triggered = true;
 				return false;
 			}
 			catch (Exception ex)
 			{
-				LogFtpMessage($"FTP[{cycleStr}]: Error uploading {remotefile} : {ex.Message}", realtime);
+				LogFtpMessage($"{prefix}: Error uploading {remotefile} : {ex.Message}", realtime);
 
 				FtpAlarm.LastMessage = $"Error uploading {remotefile} : {ex.Message}";
 				FtpAlarm.Triggered = true;
@@ -12223,7 +12269,7 @@ namespace CumulusMX
 				if (ex.InnerException != null)
 				{
 					ex = Utils.GetOriginalException(ex);
-					LogFtpMessage($"FTP[{cycleStr}]: Base exception - {ex.Message}", realtime);
+					LogFtpMessage($"{prefix}: Base exception - {ex.Message}", realtime);
 				}
 
 				// Lets start again anyway! Too hard to tell if the error is recoverable
@@ -12239,7 +12285,20 @@ namespace CumulusMX
 		// Return False if the upload failed
 		private async Task<bool> UploadString(HttpClient httpclient, bool incremental, string oldest, string data, string remotefile, int cycle, bool binary = false, bool utf8 = true, bool logfile = false, int linecount = 0)
 		{
-			var prefix = cycle >= 1000 ? $"PHP[Int-{cycle-1000}]" : $"PHP[{cycle}]";
+			string prefix;
+
+			if (cycle >= 1000)
+			{
+				prefix = "PHP[Int-{(cycle - 1000)}]";
+			}
+			else if (cycle == -1)
+			{
+				prefix = "PHP[HttpFiles]";
+			}
+			else
+			{
+				prefix = "PHP[{cycle}]";
+			}
 
 			if (string.IsNullOrEmpty(data))
 			{
@@ -12452,7 +12511,6 @@ namespace CumulusMX
 
 			return false;
 		}
-
 
 		public async Task DoSingleNoaaUpload(string filename)
 		{